Do not restart scan on host reset, start advertising when re-synced.
Previously when the host reset scanning (if active prior) would be restarted automatically.
This caused errors for some applications and has been removed since the event invokes the scan
ended callback for the app to take action.

Advertising (if active prior) would be started without regard to it's previous state.
This has been corrected to only start advertising again if it was active when the reset occurred
and maintains the application provided parameters for duration and callback from the last start.